Backend/requirements.txt
ansicon==1.89.0
arrow==1.3.0
asgiref==3.8.1
blessed==1.20.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dj-database-url==2.1.0
Django==5.0.3
django-cors-headers==4.3.1
django-picklefield==3.2
django-q==1.3.9
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
idna==3.7
jinxed==1.2.1
lxml==5.2.2
markdown-it-py==3.0.0
mdurl==0.1.2
packaging==24.0
pillow==10.2.0
psycopg2==2.9.9
Pygments==2.18.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==3.5.3
requests==2.32.3
rich==13.7.1
six==1.16.0
sqlparse==0.4.4
types-python-dateutil==2.9.0.20240316
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
whitenoise==6.6.0
